move EntityPosition to its own file
