+++ hive/initialization/sample_vehicles.py
@@ -1,13 +1,12 @@
 import functools as ft
 import logging
 import random
-from typing import Callable, Tuple
+from typing import Callable
 
 from returns.result import Result, Failure, Success
 
-from hive.model.base import Base
+from hive.model.entity_position import EntityPosition
 from hive.model.roadnetwork import Link, OSMRoadNetwork
-from hive.model.roadnetwork.link import EntityPosition
 from hive.model.vehicle.vehicle import Vehicle
 from hive.runner import Environment
 from hive.state.driver_state.autonomous_driver_state.autonomous_available import AutonomousAvailable
